requirements.txt
Flask==2.2.5
geonamescache==1.1.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.3
pyodbc==5.0.1
tabulate==0.9.0
python-dotenv==1.0.1
